+++ plugins/11_TheWarWithin/localization/zhTW.lua
@@ -635,3 +635,7 @@ L['undermine_undershirt_note'] = nil
 L['undermine_clam_chowder_label'] = nil
 L['undermine_clam_chowder_location'] = nil
 L['undermine_clam_chowder_note'] = nil
+
+L['blazefeather_peacock_note'] = nil
+L['radioactive_subject_note'] = nil
+L['george_the_big_pinch_note'] = nil